+++ src/common/FinancialSummary/FundingInformation/FundingInformation.test.js
@@ -7,6 +7,8 @@ import FundingInformation from './FundingInformation';
 const TEST_DATA = {
   initialAllocation: 100,
   netTransfers: -10,
+  allocationTo: -5,
+  allocationFrom: 20,
 };
 
 const renderFundingInformation = (data = TEST_DATA, currency = 'USD') => (render(
@@ -24,5 +26,7 @@ describe('FundingInformation component', () => {
 
     expect(getByText('$100.00')).toBeDefined();
     expect(getByText('($10.00)')).toBeDefined();
+    expect(getByText('($5.00)')).toBeDefined();
+    expect(getByText('$20.00')).toBeDefined();
   });
 });