+++ src/common/utils/index.js
@@ -1,4 +1,3 @@
-export * from './calculateTotals';
 export * from './getFiscalYearsForSelect';
 export * from './fundValidation';
 export * from './getExpenseClassesForSelect';