+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* Update Fiscal year view to include more summary information. Refs UIF-262.
 * Update group view to include more summary information. Refs UIF-264.
 * Update ledger view to include more summary information. Refs UIF-263.
 * Update Budget detail view to include more summary information. Refs UIF-261.